requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cytoolz==0.11.2
decorator==5.1.1
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
inflection==0.5.1
ipfs-api==0.2.3
ipfshttpclient==0.8.0a2
itypes==1.2.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.5.1
lru-dict==1.1.7
MarkupSafe==2.0.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
numpy==1.22.3
oauthlib==3.1.1
packaging==21.2
parsimonious==0.8.1
protobuf==3.20.1
py==1.11.0
pycryptodome==3.14.1
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.1
pytz==2021.3
requests==2.26.0
retry==0.9.2
rlp==2.0.1
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
toolz==0.11.2
uritemplate==4.1.1
urllib3==1.26.7
varint==1.0.2
web3==5.29.0
websockets==9.1
yarl==1.7.2
